# Do not enable stack tracer for ZFS performance test