@behlendorf 2018-12-04
The performance penalty is larger than I would have expected.  I wouldn't be opposed to making this behavior opt-in instead.  The primary reason the stack tracer is enabled by default to to monitor worst case stack usage.  Since this is rarely an issue any more it's reasonable to enable as needed.