@tonynguien 2018-12-04
Thanks for the background information. I'll make stack tracer enabling an
opt-in for zfs.sh and zfs-test.sh.

And I will include the performance comparative numbers as well. Thanks!

On Mon, Dec 3, 2018 at 6:39 PM Brian Behlendorf <notifications@github.com>
wrote:

> The performance penalty is larger than I would have expected. I wouldn't
> be opposed to making this behavior opt-in instead. The primary reason the
> stack tracer is enabled by default to to monitor worst case stack usage.
> Since this is rarely an issue any more it's reasonable to enable as needed.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/zfsonlinux/zfs/pull/8173#issuecomment-443937219>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ACFa_K0u9eyEFKjHkkCzbKEMv4_O8PZ2ks5u1dJkgaJpZM4Y_pVn>
> .
>