@tonynguien 2018-12-04
Thanks for the feedback. I'll make stack tracer enabling an opt-in for both
zfs.sh and zfs-test.sh.

On Mon, Dec 3, 2018 at 5:09 PM bunder2015 <notifications@github.com> wrote:

> While I've never run this set of tests myself, having these stacks might
> be beneficial to resolving issues, can we possibly make it a toggle instead
> of a hard disable?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/zfsonlinux/zfs/pull/8173#issuecomment-443918958>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ACFa_Al5FgAUcWWTdFcbXuA98eKgeyidks5u1b0igaJpZM4Y_pVn>
> .
>