@behlendorf 2018-12-04
> Should I update the PR's description/motivation text to reflect the change?

Yes, that would be helpful.  Thanks.