@tonynguien 2018-12-06
> @tonynguien what do you think about using `-S` for consistency. Settling that small nit is all that's holding up this PR.

Yup, latest commit has that change.